requirements.txt
Flask
mysql.connector-python
Flask-Cors
requests
psutil
pyinstaller